// comunicacionesDirectas/src/shared/repositories/autogenerado/wsMensajeriaClient.ts
export interface Conversacion {
  idConversacionEncriptado?: string | undefined;
  titulo?: string | undefined;
  nombreParticipante?: string | undefined;
  ultimoMensaje?: string | undefined;
  fechaUltimoMensaje?: Date;
  estadoUltimoMensaje?: number;
  tipoUltimoMensaje?: number;
  esPropioMensaje?: boolean;
  mensajesSinLeer?: number;
}